# docs/swagger.yaml
definitions:
  model.ApodQueryInput:
    properties:
      count:
        type: integer
      date:
        type: string
      end_date:
        type: string
      start_date:
        type: string
      thumbs:
        type: boolean
    type: object
  model.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  model.ListNasa:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Nasa'
        type: array
    type: object
  model.Nasa:
    properties:
      copyright:
        type: string
      date:
        type: string
      explanation:
        type: string
      hdurl:
        type: string
      id:
        type: integer
      media_type:
        type: string
      service_version:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
  description: NASA daily imaging service
  title: APOD Service
paths:
  /album/:
    post:
      consumes:
      - application/json
      description: create new albums
      parameters:
      - description: NASA
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Nasa'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: createPerm
      tags:
      - images
  /album/dt:
    get:
      consumes:
      - application/json
      description: gets image by date
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Nasa'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: getByDate
      tags:
      - images
  /album/filter:
    get:
      consumes:
      - application/json
      description: gets all images with filter
      parameters:
      - description: Filter
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.ApodQueryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListNasa'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: getWithFilter
      tags:
      - images
  /album/images:
    get:
      consumes:
      - application/json
      description: gets all images from DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListNasa'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: getImagesFromDB
      tags:
      - images
swagger: "2.0"